Draw tracker, vehicle and friend markers with the shared marker
