+++ web/src/lib/i18n/en-US.json
@@ -29,6 +29,8 @@
     "loading": "Loading...",
     "dateAdded": "Date Added",
     "lastModified": "Last Modified",
+    "lastSeen": "Last seen",
+    "updated": "Updated",
     "description": "Description",
     "icon": "Icon",
     "color": "Color",
