+++ web/src/lib/i18n/es-ES.json
@@ -29,6 +29,8 @@
     "loading": "Cargando...",
     "dateAdded": "Fecha añadida",
     "lastModified": "Última modificación",
+    "lastSeen": "Visto por última vez",
+    "updated": "Actualizado",
     "description": "Descripción",
     "icon": "Icono",
     "color": "Color",
