+++ web/src/lib/vue-marker.utils.ts
@@ -6,9 +6,11 @@ import { i18n } from '@/lib/i18n'
  *
  * Each marker is its own tiny Vue app, mounted outside the main one, so it
  * inherits none of its plugins. i18n has to be installed explicitly or any
- * marker calling `useI18n()` throws on setup and never renders — which is
- * what happened to tracker markers. Pinia needs no such wiring: stores
- * resolve through the globally active instance.
+ * marker calling `useI18n()` throws on setup — and the throw escapes into
+ * whatever was iterating markers, which is how owning a single tracker once
+ * took out every initializer queued behind the marker layers: bookmarks,
+ * notes, timeline and the portolan ribbons all silently vanished. Pinia needs
+ * no such wiring: stores resolve through the globally active instance.
  */
 export function createVueMarkerElement(
   component: Component,
@@ -19,14 +21,6 @@ export function createVueMarkerElement(
 
   // Create a Vue app instance with the component
   const app: App = createApp(component, props)
-  app.use(i18n)
-
-  // A marker is its own app, so it inherits nothing from the main one: any
-  // component calling useI18n() throws on mount, and the throw escapes into
-  // whatever was iterating markers. That killed every initializer queued
-  // after the marker layers in onStyleLoad — bookmarks, notes, timeline and
-  // the portolan ribbons all silently vanished for any account that owned a
-  // tracker. Markers need the same translations as the rest of the app.
   app.use(i18n)
 
   // Mount the app to the container
